refactor(domain): move error to separate model file
